LayoutTests/js/dom/modules/module-load-event-with-src-expected.txt

Test script.onload will be fired when the module tag with src loads and executes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
Executing a module with src attribute.
PASS moduleExecuted is true
PASS successfullyParsed is true

TEST COMPLETE
